backend/requirements.txt
langchain>=0.3
langgraph>=1.0
langchain-openai>=0.3
fastapi>=0.115
uvicorn>=0.34
websockets>=14.0
httpx>=0.28
markdownify>=0.14
aiofiles>=24.1
pydantic>=2.10
pyyaml>=6.0
python-dotenv>=1.0
